requirements.txt
groq==0.18.0
python-dotenv==1.0.0
jsonschema==4.22.0
pytest
